:bug: resolve bug auth
